requirements.txt
deta==1.1.0
numpy==1.24.3
pandas==2.0.2
plotly==5.15.0
python_dateutil==2.8.2
streamlit==1.24.0
streamlit_authenticator==0.2.2
